requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.2.0
python-magic==0.4.27
psycopg2-binary==2.9.9
django-filter==23.5
